Separate out MethodComponentContext from KNNMethodContext
