+++ src/main/java/com/amazon/opendistroforelasticsearch/knn/index/KNNMethodContext.java
@@ -57,8 +57,17 @@ public class KNNMethodContext implements ToXContentFragment {
     public static final Integer MIN_MINIMUM_DATAPOINTS = 0;
 
     public static final KNNMethodContext DEFAULT = new KNNMethodContext(KNNEngine.DEFAULT, SpaceType.DEFAULT,
-                        new KNNMethodContext.MethodComponentContext(METHOD_HNSW, Collections.emptyMap()), null, null,
+                        new MethodComponentContext(METHOD_HNSW, Collections.emptyMap()), null, null,
     KNNMethodContext.MIN_TRAINING_DATASET_SIZE_LIMIT +1, KNNMethodContext.MIN_MINIMUM_DATAPOINTS + 1);
+
+    private KNNEngine knnEngine;
+    private SpaceType spaceType;
+    private MethodComponentContext methodComponent;
+    private KNNMethodContext coarseQuantizerContext;
+    private MethodComponentContext encoder;
+
+    private Integer trainingDatasetSizeLimit;
+    private Integer minimumDatapoints;
 
     /**
      * Constructor
@@ -83,15 +92,6 @@ public class KNNMethodContext implements ToXContentFragment {
         this.trainingDatasetSizeLimit = trainingDatasetSizeLimit;
         this.minimumDatapoints = minimumDatapoints;
     }
-
-    private KNNEngine knnEngine;
-    private SpaceType spaceType;
-    private MethodComponentContext methodComponent;
-    private KNNMethodContext coarseQuantizerContext;
-    private MethodComponentContext encoder;
-
-    private Integer trainingDatasetSizeLimit;
-    private Integer minimumDatapoints;
 
     /**
      * Gets the main method component
@@ -343,92 +343,5 @@ public class KNNMethodContext implements ToXContentFragment {
         }
 
         return builder;
-    }
-
-    /**
-     * MethodComponentContext represents a single building block of a knn library index.
-     *
-     * Each component is composed of a name and a map of parameters.
-     */
-    public static class MethodComponentContext implements ToXContentFragment {
-        public MethodComponentContext(String name, Map<String, Object> parameters) {
-            this.name = name;
-            this.parameters = parameters;
-        }
-
-        private String name;
-        private Map<String, Object> parameters;
-
-        /**
-         * Parses the object into MethodComponentContext
-         *
-         * @param in Object to be parsed
-         * @return MethodComponentContext
-         */
-        public static MethodComponentContext parse(Object in) {
-            if (!(in instanceof Map)) {
-                throw new MapperParsingException("Unable to parse MethodComponent");
-            }
-
-            @SuppressWarnings("unchecked")
-            Map<String, Object> methodMap = (Map<String, Object>) in;
-            String name = "";
-            Map<String, Object> parameters = null;
-
-            String key;
-            Object value;
-            for (Map.Entry<String, Object> methodEntry : methodMap.entrySet()) {
-                key = methodEntry.getKey();
-                value = methodEntry.getValue();
-                if (NAME.equals(key)) {
-                    if (!(value instanceof String)) {
-                        throw new MapperParsingException("Component name should be a string");
-                    }
-
-                    name = (String) value;
-                } else if (PARAMETERS.equals(key)) {
-                    if (!(value instanceof Map)) {
-                        throw new MapperParsingException("Unable to parse parameters for main method component");
-                    }
-
-                    @SuppressWarnings("unchecked")
-                    Map<String, Object> parameters1 = (Map<String, Object>) value;
-                    parameters = parameters1;
-                } else {
-                    throw new MapperParsingException("Invalid parameter for MethodComponentContext: " + key);
-                }
-            }
-
-            if (name.isEmpty()) {
-                throw new MapperParsingException(NAME + " needs to be set");
-            }
-
-            return new MethodComponentContext(name, parameters);
-        }
-
-        @Override
-        public XContentBuilder toXContent(XContentBuilder builder, Params params) throws IOException {
-            builder.field(NAME, name);
-            builder.field(PARAMETERS, parameters);
-            return builder;
-        }
-
-        /**
-         * Gets the name of the component
-         *
-         * @return name
-         */
-        public String getName() {
-            return name;
-        }
-
-        /**
-         * Gets the parameters of the component
-         *
-         * @return parameters
-         */
-        public Map<String, Object> getParameters() {
-            return parameters;
-        }
     }
 }
